update: added support for setting l1_da_mode from chain config
