+++ configs/presets/integration.yaml
@@ -1,5 +1,6 @@
 chain_name: "Starknet Sepolia"
 chain_id: "SN_INTEGRATION_SEPOLIA"
+l1_da_mode: "Calldata"
 feeder_gateway_url: "https://integration-sepolia.starknet.io/feeder_gateway/"
 gateway_url: "https://integration-sepolia.starknet.io/gateway/"
 native_fee_token_address: "0x04718f5a0fc34cc1af16a1cdee98ffb20c31f5cd61d6ab07201858f4287c938d"
