+++ madara/crates/primitives/chain_config/Cargo.toml
@@ -22,6 +22,7 @@ starknet_api.workspace = true
 
 # Madara
 mp-utils.workspace = true
+mp-rpc = { workspace = true }
 
 # Other
 anyhow.workspace = true
