+++ madara/crates/primitives/chain_config/src/lib.rs
@@ -1,7 +1,9 @@
 mod chain_config;
+mod l1_da_mode;
 mod rpc_version;
 mod starknet_version;
 
 pub use chain_config::*;
+pub use l1_da_mode::*;
 pub use rpc_version::*;
 pub use starknet_version::*;
